# api/swagger.yaml
...
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "FC Api"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "agnetiuslee@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/ronaldSengkey/fitClub/api/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /bank/{token}:
    get:
      tags:
      - "coach"
      summary: "Registration access"
      description: ""
      operationId: "getBank" 
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "param"
        in: "header"
        description: "token of member device"
        required: true
        type: "string"
      - name: "token"
        in: "path"
        description: "token of member device"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "transaction"
...
  /transaction/request/{param}/{token}:
      get:
        tags:
        - "transaction"
        summary: "Registration access"
        description: ""
        operationId: "transactionRequest"
        consumes:
        - "multipart/form-data"
        - "application/x-www-form-urlencoded"
        - "application/json"
        produces:
        - "application/xml"
        - "application/json"
        parameters:
        - name: "token"
          in: "path"
          required: true
          type: "string"
        - name: "param"
          in: "path"
          required: true
          type: "string"
        responses:
          200:
            description: "Success"
          500:
            description: "Internal error"
          401:
            description: "Unauthorize"
          404:
            description: "Data not found"
        x-swagger-router-controller: "transaction"
...
  /member/payment/{token}:
      post:
        tags:
        - "transaction"
        summary: "Registration access"
        description: ""
        operationId: "memberPayment"
        consumes:
        - "multipart/form-data"
        - "application/x-www-form-urlencoded"
        - "application/json"
        produces:
        - "application/xml"
        - "application/json"
        parameters:
        - name: "token"
          in: "path"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "data credential guest"
          required: true
          schema:
            $ref: "#/definitions/paymentParam"
        responses:
          200:
            description: "Success"
          500:
            description: "Internal error"
          401:
            description: "Unauthorize"
          404:
            description: "Data not found"
        x-swagger-router-controller: "transaction"
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  paymentParam:
    type: "object"
    properties:
      nominal:
        type: "string"
      memberCategory: #cateogry member gold / silver / platinum
        type: "string"
      placeId: #id tempat mendaftar
        type: "string"
      paymentType: #cash / transfer
        type: "string"
      description: #keterangan
        type: "string"
...